CODE/Stage_2/FaceTracking_old/requirements.txt
opencv-python
numpy
pillow
torch
torchvision
tqdm
scikit-learn
pandas
mtcnn
deepface
hdbscan
matplotlib
imutils
umap-learn